TypeScript tracks-mixin
