+++ src/js/types/generic.type.ts
@@ -23,7 +23,19 @@ export type DrmConfig = GenericObject;
 export type PluginObj = GenericObject;
 
 export type TextTrackLike = TextTrack & {
-    _id: string;
+    _id?: string;
     data: Array<any>;
     name: string;
+    inuse?: boolean;
+    embedded?: boolean;
+    sideloaded?: boolean;
+    groupid?: string; // Shaka-player 'subs' tracks
+    source?: 'mpegts'; // Legacy Flash provider attribute
+};
+
+export type MetadataEvent = {
+    metadataTime?: number;
+    metadataType?: string;
+    metadata: GenericObject;
+    programDateTime?: string;
 };
